Organize classes into suitable java packages
