+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-      Boooo! I'm DaveTheBrave
+      Boooo! I'm davethebrave.DaveTheBrave
       What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
